Backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.12.0
psycopg2-binary>=2.9.10
ultralytics==8.0.227
opencv-python==4.8.1.78
celery==5.3.4
redis==5.0.1
Pillow>=9.0
numpy>=1.24.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
